Add sidekiq, redis, and create the first worker.
