+++ Gemfile
@@ -17,7 +17,11 @@ gem "pry"
 gem "pry-rails"
 gem "rails", "~> 6.1"
 gem "rollbar"
+gem 'redis'
+gem 'redis-rails'
+gem 'redis-store'
 gem "mimemagic"
+gem "sidekiq"
 
 group :development, :test do
   gem "pry-remote"
