+++ config/routes.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
-
+require "sidekiq/web"
 
 Rails.application.routes.draw do
+  mount Sidekiq::Web => "/sidekiq"
 end
